fix relative number toggling
number and relativenumber aren't different states that oppose each
other, but two things that can both be happening. I always want number,
but relativenumber needs to be set and unset (or set no…, as it happens).

I'm not sure if this was working before and stopped, or if it was always
a problem. It's been a while.