switch from Vundle to Plug
There are a few arguments for this. Mainly, I like that Plug is
continuing to be worked on, and it's a bit simpler to deal with. No more
submdules, eh?

Also, this is an excellent time to update some things, like the ctrl-p
plugin. And get rid of the Puppet one.