move color-setting to a separate config file
The reason the color scheme wasn't working should've been obvious: gui
is sorted before window, so the gui color scheme was set, and then it
was overwritten later. This is a lesson in not having the same thing get
set in multiple files.

Split off color stuff to a separate file, with a conditional in it to
set the color scheme. For some reaosn, though, it's loading the normal
Tomorrow-Night theme, not my tweaked version with visible line numbers.
Something to figure out.