split NERDCommenter settings to separate file
Going with `commenting.vim`, because I don't need these files to
correspond exactly with the plugin that's providing the functionality.

For ctrlp, sure, I'm going to use that (though my settings file handles
the ctrlp plugin and the related -funky one). Same with gitgutter, at
least so far (we'll see what the future brings). But for commenting and
the statusline, I figure the files can be just "commenting" and
"statusline", not pointing out what I'm using.

That's maybe a bit more than needed to be said on a commit like this.

But one thing to point out is I was going to change the `map`s to
`noremap`s, and those didn't work. Maybe because of the `<plug>` stuff,
which I don't fully understand yet.