organize editing settings
Isn't it weird to have a settings file called `editing`? I mean, this is
an editor. But not everything done in it is editing. This is for the
editing-based stuff. You get it.

Anyway, putting stuff in groups, so I can see what's going on here. And
moving out the stuff that's not editing. For example, I think syntax
highlighting makes more sense with the filetypes than it does here.

Also, making a note to really figure out the deal with autoindentation.
I want that part of the file either gone, or modified to not say it
needs to be fixed.