add mappings for quick-extra settings
If something strikes my fancy while I'm working, I'd like to quickly add
it to the config and deal with putting it in the right spot later. But I
want to do it in a controlled way, so I have a special file for this
quick extra (or x-tra) settings.