fix "broken" vim-config filetype setting
It wasn't broken at all, but it's just the way I have my machine set up.
When I'm directly in the (symlinked) directory, it doesn't think I'm in
~/.vim, but actually in the real directory. So my quick fix now is just
putting the full path to that directory. Maybe later I can fix this so
it see the pattern as I want it to.