only one xtra-edit window at a time
This is a bit of a monster, to keep only a single "xtra" config edit
window open at once, what with every WinEnter being checked to see if
"config/xtra.vim" is still being edited. But it's what I could figure
out that works.

Important notes for my (personal) development:

- There doesn't seem to be a WinClose or WinQuit kind of event
- WinLeave happens *before* the window is left
- I couldn't get a regex (/\v\<config\/xtra\.vim$/) to work, hence the
  split-slice-join