use the git tags
I had this whole thing where I was trying to add the tags from the git
dir no matter where I was in a repo directory structure (if I was in a
git repo, that is), but it wasn't working. For one, the file was being
loaded before fugitive, which wasn't something I was easily finding a
way to deal with. And even if I re-sourced so that fugitive was loaded,
it wasn't getting the tag file I wanted.

Forget it. This is so much easier and straightforward.