restoring fzf git status
`GFiles?` isn't a separate command, but an argument to GFiles. And I've
replaced that with a proximity-aware version that doesn't know what to
do with an argument.

I don't know that much about how to handle args and bangs with commands,
so I've just made a new command to replace GFiles? when appropriate. And
it doesn't even have proximity handling, because I'm not sure best how
to handle both the proximity handling and the diff preview. (I get the
diff preview like this, and if I pipe the status output through `cut`, I
can get the proximity handling but then the preview is of the file, not
the diff.)