hide more parts of statusline, for blame view
This took some doing, putting in a bunch of component functions to
override the normal components and check the filetype to decide whether
or not to return the empty string. There's some possibility of config
with component visibility, but there's not a lot of documentation around
that and it seems like the lightline author really wants component
functions to be used over the visibility checking.

The lineinfo and percent were especially annoying. That's where I would
_really_ like to just use a function to decide whether or not to show
the data, not have to figure out the data myself. Or be able to return
a string like '%3p%%' and it would Just Work™. Or different lightline
configs based on filetype?

I'll look more into this another time. At least now I've got it looking
right again.