NERDTree: find the current file if possible
This just makes my own NERDTreeToggle with a maybe find in there.

I couldn't get the on-demand loading for plug working correctly, so I
guess nerdtree is just loaded in general now.

Also, moving nerdtree config to its own file, but not naming it directly
after the plugin, but the general idea.