save test commands in repo
Instead of doing these entirely in a project-specific .vimrc, the
entire binding can happen here in this repo, and the project-specific
.vimrc can just define the necessary variable.

Maybe later I'll move to something buffer-specific and check if the
variable is even defined, because not every buffer makes sense to run
the same test command, or tests at all. Something to consider for later.